requirements.txt
ultralytics>=8.1.0
opencv-python>=4.9.0
numpy>=1.24
scipy>=1.11
pandas>=2.1
pyyaml>=6.0
rich>=13.7
loguru>=0.7
lapx>=0.5.5
torch>=2.1
torchvision>=0.16
tqdm>=4.66
paddleocr>=2.7.0.3
easyocr>=1.7.1
rapidfuzz>=3.6
onnx>=1.16
onnxruntime>=1.18
fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart>=0.0.9
starlette>=0.37
streamlit>=1.36
pydantic>=2.7
shapely>=2.0
